phase 6, tests not finished

+++ shared/test-utils/src/commonMain/kotlin/com/scanium/test/Matchers.kt
@@ -198,7 +198,7 @@ fun assertTrackerStats(
     stats: TrackerStats,
     expectedActiveCandidates: Int? = null,
     expectedConfirmedCandidates: Int? = null,
-    expectedCurrentFrame: Int? = null
+    expectedCurrentFrame: Long? = null
 ) {
     expectedActiveCandidates?.let {
         assertEquals(
